+++ src/main/java/com/microsoft/sqlserver/jdbc/dataclassification/ColumnSensitivity.java
@@ -1,4 +1,4 @@
-package com.microsoft.sqlserver.jdbc;
+package com.microsoft.sqlserver.jdbc.dataclassification;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,11 +6,11 @@ import java.util.List;
 public class ColumnSensitivity {
     public List<SensitivityProperty> sensitivityProperties;
 
-	public ColumnSensitivity(List<SensitivityProperty> sensitivityProperties) {
+    public ColumnSensitivity(List<SensitivityProperty> sensitivityProperties) {
         this.sensitivityProperties = new ArrayList<SensitivityProperty>(sensitivityProperties);
     }
 
     public List<SensitivityProperty> getSensitivityProperties() {
-		return sensitivityProperties;
-	}
+        return sensitivityProperties;
+    }
 }
