+++ src/main/java/com/microsoft/sqlserver/jdbc/dataclassification/Label.java
@@ -1,0 +1,20 @@
+package com.microsoft.sqlserver.jdbc.dataclassification;
+
+public class Label {
+    public String name;
+    public String id;
+
+    public Label(String name,
+            String id) {
+        this.name = name;
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getId() {
+        return id;
+    }
+}
